Refactor: all hybrid globalization js methods in one file.
