+++ src/mono/wasm/runtime/net6-legacy/hybrid-globalization.ts
@@ -3,9 +3,71 @@
 
 import { Module } from "../imports";
 import { mono_wasm_new_external_root } from "../roots";
-import {MonoString } from "../types";
+import {MonoString, MonoStringRef } from "../types";
 import { Int32Ptr } from "../types/emscripten";
-import { js_string_to_mono_string_root } from "../strings";
+import { conv_string_root, js_string_to_mono_string_root } from "../strings";
+import { setU16 } from "../memory";
+
+export function mono_wasm_change_case_invariant(exceptionMessage: Int32Ptr, src: number, srcLength: number, dst: number, dstLength: number, toUpper: number) : void{
+    try{
+        const input = get_utf16_string(src, srcLength);
+        let result = toUpper ? input.toUpperCase() : input.toLowerCase();
+        // Unicode defines some codepoints which expand into multiple codepoints,
+        // originally we do not support this expansion
+        if (result.length > dstLength)
+            result = input;
+
+        for (let i = 0; i < result.length; i++)
+            setU16(dst + i*2, result.charCodeAt(i));
+    }
+    catch (ex: any) {
+        pass_exception_details(ex, exceptionMessage);
+    }
+}
+
+export function mono_wasm_change_case(exceptionMessage: Int32Ptr, culture: MonoStringRef, src: number, srcLength: number, dst: number, destLength: number, toUpper: number) : void{
+    const cultureRoot = mono_wasm_new_external_root<MonoString>(culture);
+    try{
+        const cultureName = conv_string_root(cultureRoot);
+        if (!cultureName)
+            throw new Error("Cannot change case, the culture name is null.");
+        const input = get_utf16_string(src, srcLength);
+        let result = toUpper ? input.toLocaleUpperCase(cultureName) : input.toLocaleLowerCase(cultureName);
+        if (result.length > destLength)
+            result = input;
+
+        for (let i = 0; i < destLength; i++)
+            setU16(dst + i*2, result.charCodeAt(i));
+    }
+    catch (ex: any) {
+        pass_exception_details(ex, exceptionMessage);
+    }
+    finally {
+        cultureRoot.release();
+    }
+}
+
+export function mono_wasm_compare_string(exceptionMessage: Int32Ptr, culture: MonoStringRef, str1: number, str1Length: number, str2: number, str2Length: number, options: number) : number{
+    const cultureRoot = mono_wasm_new_external_root<MonoString>(culture);
+    try{
+        const cultureName = conv_string_root(cultureRoot);
+        const string1  = get_utf16_string(str1, str1Length);
+        const string2 = get_utf16_string(str2, str2Length);
+        const casePicker = (options & 0x1f);
+        const locale = cultureName ? cultureName : undefined;
+        const result = compare_strings(string1, string2, locale, casePicker);
+        if (result == -2)
+            throw new Error("$Invalid comparison option.");
+        return result;
+    }
+    catch (ex: any) {
+        pass_exception_details(ex, exceptionMessage);
+        return -2;
+    }
+    finally {
+        cultureRoot.release();
+    }
+}
 
 export function get_utf16_string(ptr: number, length: number): string{
     const view = new Uint16Array(Module.HEAPU16.buffer, ptr, length);
